Move submit resource to admin::rest_api::actix::submit module
