+++ libsplinter/src/admin/rest_api/actix/mod.rs
@@ -16,3 +16,4 @@
 pub(super) mod proposals_read;
 #[cfg(feature = "proposal-read")]
 pub(super) mod proposals_read_circuit_id;
+pub(super) mod submit;
